+++ CHANGELOG.md
@@ -12,7 +12,9 @@ Create: Mobile Packages - v0.7.7 - 1.21.1 - unreleased
 - Fix Playernetworks Screen now showing network connection status: Member / Owner (#346)
 - Fix NetworkSettingsScreen scrollbar not working (#328)
 - Fix Lock / Unlock button not updating (#331)
+- Add `/cmp network rename` admin command to rename networks from console
 - PlayerNetworksScreen now shows join/leave button depending on membership
+- NetworkSettingsScreen now shows the network UUID (shortened) at the bottom of the screen
 
 ------------------------------------------------------
 Create: Mobile Packages - v0.7.6 - 1.21.1 - 08.05.2026